cambios por que me apuran

+++ tp2/estudiantes/src/main/java/aed/Edr.java
@@ -242,7 +242,7 @@ public class Edr {
     
     public int[] chequearCopias() {
     
-        double[][] cantDeRtasAPreg = new int[10][_solCanonica.length];   // O(R) porque creamos R arrays con 10 posiciones
+        double[][] cantDeRtasAPreg = new double[10][_solCanonica.length];   // O(R) porque creamos R arrays con 10 posiciones
         
         // TODO: Resolver duda: habría que inicializar en 0?
